requirements.txt
# Development tools
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.3.3

# Documentation
Sphinx==7.2.5
sphinx-rtd-theme==1.3.0

# Environment management
python-dotenv==1.0.0

# Packages forgot to add will organize later
Flask==3.1.0
flask-cors==5.0.1
azure-storage-blob==12.25.1
openai==1.75.0
Werkzeug==3.1.3  
azure-core==1.33.0  
requests==2.32.3  

#database packages

#Flask-SQLAlchemy>=3.0.0
#Flask-Migrate>=4.1.0
#SQLAlchemy>=2.0.40
#alembic>=1.10.0

#database new req2.0
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9